/ppbox/branches/1.2.0 /src/ppbox 修正一些文件不存在的 include emailto: box-dev
git-svn-id: http://svn01/repositories/r01/ppbox/branches/1.2.0@6774 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ MuxerBase.cpp
@@ -245,7 +245,7 @@ namespace ppbox
         }
 
         void MuxerBase::format_type(
-            boost::uint8_t t)
+            boost::uint32_t t)
         {
             format_type_ = t;
         }
